User ID & password
login admin, the password is admin and select user type admin, then on view customers, it shows the user name and a/c no. 
if you want to log in as a customer opens an account or log in existing customer.
user  - a/c
password - user name
user type - customer

Thanks.